third_party/WebKit/LayoutTests/platform/win/cssom/ahem-ex-units-expected.txt

Test ex unit with Ahem font

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL parseFloat(computedStyleDiv("word-spacing", "1ex")) should be within 0.05 of 12.8. Was 13.008.
FAIL parseFloat(computedStyleRect("word-spacing", "1ex")) should be within 0.05 of 12.8. Was 13.008.
PASS successfullyParsed is true

TEST COMPLETE
